Enable torch version dependent compilation of record_module and iter_params (#7362)
Dynamo breaks graphs because currently compile is disabled for a number
of functions such as `iter_params` and `record_module`.

The above functions compile successfully for at least PyTorch version
2.7.0.

We enable the compilation based on the user PyTorch version using a new
`compiler.enable(min_version=None)` decorator.
This should avoid the corresponding graph breaks and improve the
performance.

---------

Signed-off-by: Max Kovalenko <mkovalenko@habana.ai>
Co-authored-by: Masahiro Tanaka <81312776+tohtana@users.noreply.github.com>
Co-authored-by: Logan Adams <114770087+loadams@users.noreply.github.com>