Take **kwargs in __init__ of DeepSpeedZeroOptimizer subclasses (#7634)
DeepSpeedZeroOptimizer provides a rich, evolving list of keyword
arguments. It is tedious and error-prone to list all of them in its
subclasses. As an example, the recent introduction of zenflow_config in
the middle of that list has caused unit test failures (e.g.
https://github.com/deepspeedai/DeepSpeed/actions/runs/18560070656/job/52906645682?pr=7633)

Convert the keyword argument list in DeepSpeedZeroOptimizer subclasses
to **kwargs for the consistency of configurable items and their default
values. Passing an unknown parameter to such subclasses will now raise
an error on their call to DeepSpeedZeroOptimizer.__init__() instead of
their own __init__(). It still ensures that typo in such parameters fail
early.

Signed-off-by: Junjie Mao <junjie.mao@linux.alibaba.com>